requirements.txt
flask
python-dotenv
elevenlabs
schedule
supabase